+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 **ATTN**: This project uses [semantic versioning](http://semver.org/).
 
+## 0.2.1 - 2019-10-22
+- Fix empty output, return 204 "no content"
+- use struc interface between srv & wks
+
+## 0.2.0 - 2019-10-12
+- Add polling support for long tasks
+
+## 0.1.3 - 2019-06-19
+- Fix logrotate file loc
+- Add icon
+- Add windows file info
+- Remove gonfig lib
 
 ## 0.1.2 - 2019-05-18
 Move to opensource, using AGPLv3.